% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{australia10}
\alias{australia10}
\title{Australian mortality data, 2010}
\format{A data frame with 3,322 rows and 3 variables:
\describe{
\item{cause}{ICD-10 code representing cause of death}
\item{sex}{Gender}
\item{deaths}{Number of deaths}
}}
\usage{
australia10
}
\description{
A dataset containing australian mortality data, obtained from Stata version 15.
}
\note{
The R code used to download and process the dataset from Stata is available \href{https://raw.githubusercontent.com/ellessenne/comorbidity/master/data-raw/make-data.R}{here}.
}
\keyword{datasets}
