% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_algorithms.R
\name{available_algorithms}
\alias{available_algorithms}
\title{Display Currently Supported Algorithms}
\usage{
available_algorithms()
}
\description{
This function prints all (currently) supported and implemented
comorbidity mapping, and for each one of those, each supported scoring and
weighting algorithm.
}
\examples{
available_algorithms()
}
