% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidityAnalysis.R
\name{comorbidityAnalysis}
\alias{comorbidityAnalysis}
\title{Comorbidity Analysis \code{cAnalysis}}
\usage{
comorbidityAnalysis(input, codesPth, databasePth, ageRange = c(0, 100),
  sex = "ALL", score, correctedPval = 1, correctionMethod = "fdr",
  oddsRatio, rr, phi, cores = 1, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{A comorbidity object, obtained with the query function.}

\item{codesPth}{Determines the path where the file with the index diseases is 
located (indexDiseaseCode)}

\item{databasePth}{Determines the path where the three required input files 
(patientData, diagnosisData, admissionData) are located.}

\item{ageRange}{Determines what is the age range of interest for
performing the comorbidity analysis. By default it is set from 0 to 100 
years old.}

\item{sex}{Determine what is the sex of interest for 
performing the comorbidity analysis. By default \code{ALL}. Change it to the 
sex of interest for your comorbidity analysis.}

\item{score}{The comorbidity score is a measure based on  the observed comorbidities
and the expected ones, based on the occurrence of each disease.}

\item{correctedPval}{By default 1.}

\item{correctionMethod}{A Fisher exact test for each pair of diseases is performed to assess 
the null hypothesis of independence between the two diseases. The Benjamini-Hochberg 
false discovery rate method ("fdr") is applied to correct for multiple testing by default. 
However user can select the best correction method for the analysis. The adjustment methods 
include the Bonferroni correction ("bonferroni"),  Holm correction ("holm"), Hochberg correction
 ("hochberg"), Hommel ("hommel") and Benjamini & Yekutieli ("BY").}

\item{oddsRatio}{The odds ratio represents the increased chance that someone 
suffering disease X will have the comorbid disorder Y.}

\item{rr}{The relative risk refers to the fraction between the number of 
patients diagnosed with both diseases and random expectation based on disease 
prevalence.}

\item{phi}{The Pearsons correlation for binary variables (Phi) measures the 
robustness of the comorbidity association.}

\item{cores}{By default \code{1}. To run parallel computations on machines 
with multiple cores or CPUs, the cores argument can be changed.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{cAnalysis}
}
\description{
Given an object of type \code{comoRbidity}, a comorbidity analysis is perform, 
for the subset of population under specific conditions of age and sex. It 
generates a \code{cAnalysis} object.
}
\examples{
load(system.file("extdata", "comorbidity.RData", package="comoRbidity"))
ex1 <- comorbidityAnalysis( 
              input              = comor_obj,
              databasePth      = system.file("extdata", package="comoRbidity"),
              codesPth         = system.file("extdata", package="comoRbidity"),
              ageRange         = c(0,50),
              sex              = "Female", 
              score            = 1,
              correctionMethod = "fdr", 
              correctedPval  = 1
              )
}
