% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidityAnalysisMolecular.R
\name{comorbidityAnalysisMolecular}
\alias{comorbidityAnalysisMolecular}
\title{Comorbidity Analysis \code{comorbidityAnalysisMolecular}}
\usage{
comorbidityAnalysisMolecular(input, pValue = FALSE, nboot = 100,
  ncores = 1, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{A \code{molecularComorbidity} object, obtained with the 
\code{queryMolecular} function.}

\item{pValue}{Determines if the p-value is estimated or not. By default it is 
set to \code{'FALSE'}. The \code{pValue} argument can be set to \code{'TRUE'} 
in order to estimate the P-value associated to each Jaccard Index.}

\item{nboot}{Determines the number of random times that the Jaccard Index is 
computed using random sets. By default it is set to \code{100}. The value 
of the argument can be changed to any other numeric variable.}

\item{ncores}{By default \code{1}. To run parallel computations on machines 
with multiple cores or CPUs, the \code{ncores} argument can be changed.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{molecularcAnalysis}
}
\description{
Given an object of type \code{molecularComorbidity}, a comorbidity analysis is perform, 
taking into account the number of genes shared between the disorders.
It generates a \code{molecularcAnalysis} object.
}
\examples{
load(system.file("extdata", "mc.RData", package="comoRbidity"))
ex1 <- comorbidityAnalysisMolecular( 
              input      = mc, 
              pValue     = FALSE,
              verbose    = FALSE
              )
}

