% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2molecularcomorbidity.R
\name{dataframe2molecularcomorbidity}
\alias{dataframe2molecularcomorbidity}
\title{From a file containing a table with gene and disease information generates a \code{molecularComorbidity}
object for comorbidity analysis}
\usage{
dataframe2molecularcomorbidity(filePth, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{filePth}{The file name and with the complete path where the file with
the genes and disorders of interest is located}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{molecularComorbidity}
}
\description{
Given a file with gene and diseases, the \code{dataframe2molecularcomorbidity}
generates a \code{molecularComorbidity} object.
}
\examples{
ex1 <- dataframe2molecularcomorbidity(
              filePth = system.file("extdata", "genediseaseTable.txt", package="comoRbidity"),
              verbose  = FALSE
              )
}
