% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeUse.R
\name{diagnosticUse}
\alias{diagnosticUse}
\title{A graphical summary of the code uses.}
\usage{
diagnosticUse(codesPth, input, barColor = "darkblue", cutOff = 0,
  interactive = FALSE)
}
\arguments{
\item{codesPth}{Determines the path where the file with the index diseases is 
located (indexDiseaseCode)}

\item{input}{A comorbidity object, obtained applying query function.}

\item{barColor}{Determines the bar color. By default \code{darkblue}.}

\item{cutOff}{By default \code{0}. Change it to show only those codes with a 
percentage of use equal or higher than the cutOff value.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A barplot with the diagnosis code uses.
}
\description{
Given an object of class \code{comorbidity}, and the file which contains the
codes.
}
\examples{
load(system.file("extdata", "comorbidity.RData", package="comoRbidity"))
diagnosticUse( codesPth    = system.file("extdata", package="comoRbidity"), 
               input       = comor_obj, 
               cutOff      = 0.5, 
               barColor    = "darkcyan"
           )
}

