% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasePrevalence.R
\name{diseasePrevalence}
\alias{diseasePrevalence}
\title{A graphical summary of the disease prevalence.}
\usage{
diseasePrevalence(input, maleCode = "Male", femaleCode = "Female",
  databasePth, barColor = "lightblue", verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{Object of \code{comorbidity} class.}

\item{maleCode}{Characters(s) used to determine the male condition of a patient. 
Depending on the database it can be determined, for example, as \code{Male}, .
\code{MALE}, \code{M}, with digits as \code{0} or \code{1}.}

\item{femaleCode}{Characters(s) used to determine the female condition of a patient. 
Depending on the database it can be determined, for example, as \code{Female}, .
\code{FEMALE}, \code{F}, with digits as \code{0} or \code{1}.}

\item{databasePth}{Determines the path where the three required input files 
(patientData, diagnosisData, admissionData) are located.}

\item{barColor}{Determines the bar color. By default \code{lightblue}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A barplot with disease prevalence in all the population and according to 
the gender.
}
\description{
Given an object of class \code{comorbidity}, and the characters used to 
specify the gender, a barplot showing the disease prevalence in general 
and according to gender is obtained.
}
\examples{
load(system.file("extdata", "comorbidity.RData", package="comoRbidity"))
diseasePrevalence( input      = comor_obj, 
           maleCode   = "Male", 
           femaleCode = "Female",
           databasePth      = system.file("extdata", package="comoRbidity"), 
           barColor    = "darkcyan"
           )
}

