% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIndexDisease.R
\name{extractIndexDisease}
\alias{extractIndexDisease}
\title{Obtain the index diseases from a \code{comoRbidity} object}
\usage{
extractIndexDisease(input)
}
\arguments{
\item{input}{An object of class \code{comoRbidity}}
}
\value{
A list with the index diseases present in the input data
}
\description{
Given an object of type \code{comoRbidity}, extract the index diseases 
present in the data.
}
\examples{
indexDis <- extractIndexDisease( obj )
}
