% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapDirection.R
\name{heatmapDirection}
\alias{heatmapDirection}
\title{Plot the directionality results in a heatmap.}
\usage{
heatmapDirection(input, interactive = FALSE, fromAtoBColor = "darkgreen",
  fromBtoAColor = "orange", noDirectionColor = "grey", verbose = FALSE)
}
\arguments{
\item{input}{A \code{data.frame} obtained after applying the \code{directionality}
function}

\item{interactive}{Determines if the output heatmap is interactive or not. 
By default the \code{interactive} argument is set up as \code{FALSE}. The value 
of the argument can be changed to \code{TRUE}, as a result an interactive 
heatmap will be obtained.}

\item{fromAtoBColor}{Determines the heatmap color for the direction A to B.
By default it is set to \code{"olivegreen"}.}

\item{fromBtoAColor}{Determines the heatmap color for the direction B to A.
By default it is set to \code{"orange"}.}

\item{noDirectionColor}{Determines the heatmap color when there is no preferred direction.
By default it is set to \code{"grey"}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A heatmap
}
\description{
Given an \code{dataframe} with the directionality results obtained after apply
the \code{directionality} function, a heatmap plot of the results is obtained.
}
\examples{
load(system.file("extdata", "directionalityExample.RData", package="comoRbidity"))
htmpDirection <- heatmapDirection( input            = directionalityEx, 
                                   fromAtoBColor    = "darkgreen", 
                                   fromBtoAColor    = "orange", 
                                   noDirectionColor = "grey")
htmpDirection
}
