% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexRatioHeatmap.R
\name{heatmapSexRatio}
\alias{heatmapSexRatio}
\title{Sex Ratio Heatmap}
\usage{
heatmapSexRatio(input, maleColor = "blue", femaleColor = "red",
  verbose = FALSE, warnings = TRUE, interactive = FALSE)
}
\arguments{
\item{input}{Data frame obtained after applying the \code{sexRatio} function.}

\item{maleColor}{Determines the heatmap color for those comorbidities that
are more likely in men than women. By default \code{"blue"}.}

\item{femaleColor}{Determines the heatmap color for those comorbidities that
are more likely in women than men. By default \code{"red"}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A dataframe with the sex ratio estimation is obtained
}
\description{
Given the data frame obtained after apply the \code{sexRatio} function, a 
heatmap representing the sex ratio results is obtained.
}
\examples{
load(system.file("extdata", "srEstimation.RData", package="comoRbidity"))
heatmapSexRatio( input = srEstimation )
}

