% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_2009}
\alias{icd10_2009}
\title{ICD-10 Diagnostic Codes, 2009 Version}
\format{A data frame with 10,817 rows and 4 variables:
\describe{
\item{Code}{ICD-10 diagnostic code}
\item{Code.clean}{ICD-10 diagnostic code, removing all punctuation}
\item{ICD.title}{Code description, in plain English.}
\item{Status}{Additional information, if available.}
}}
\source{
CDC Website: https://goo.gl/6e2mvb
}
\usage{
icd10_2009
}
\description{
A dataset containing the 2009 version of the ICD-10 codes.
}
\note{
The R code used to download and process the dataset from the CDC website is available \href{https://github.com/ellessenne/charlson/blob/master/data-raw/make_data.R}{here}.
}
\keyword{datasets}
