% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_2011}
\alias{icd10_2011}
\title{ICD-10 Diagnostic Codes, 2011 Version}
\format{
A data frame with 10,856 rows and 4 variables:
\describe{
\item{Code}{ICD-10 diagnostic code}
\item{Code.clean}{ICD-10 diagnostic code, removing all punctuation}
\item{ICD.title}{Code description, in plain English.}
\item{Status}{Additional information, if available.}
}
}
\source{
CDC Website: https://goo.gl/rcTJJ2
}
\usage{
icd10_2011
}
\description{
A dataset containing the 2011 version of the ICD-10 codes.
}
\note{
The R code used to download and process the dataset from the CDC website is available \href{https://raw.githubusercontent.com/ellessenne/comorbidity/master/data-raw/make-data.R}{here}.
}
\keyword{datasets}
