% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm_2017}
\alias{icd10cm_2017}
\title{ICD-10-CM Diagnostic Codes, 2017 Version}
\format{
A data frame with 71,486 rows and 2 variables:
\describe{
\item{Code}{ICD-10-CM diagnostic code}
\item{Description}{Description of each code}
}
}
\usage{
icd10cm_2017
}
\description{
A dataset containing the 2017 version of the ICD10-CM coding system.
}
\note{
The R code used to download and process the dataset from the CDC website is available \href{https://raw.githubusercontent.com/ellessenne/comorbidity/master/data-raw/make-data.R}{here}.
}
\keyword{datasets}
