% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Plot the comorbidity analysis results in a network}
\usage{
network(input, databasePth, layout = "layout.fruchterman.reingold",
  selectValue = "score", title = "Comorbidity network", cutOff = 0,
  npairs = 0, prop = 1, interactive = FALSE, diseaseColor = "pink",
  comorColor = "lightblue", verbose = FALSE)
}
\arguments{
\item{input}{A \code{cAnalysis} or \code{molecularcAnalysis} object, obtained 
by applying the \code{comorbidityAnalysis} or \code{comorbidityAnalysisMolecular} 
function}

\item{databasePth}{Determines the path where the intermediate RData objects 
have been created. It is the same path where the three required input files 
(patientData, diagnosisData, admissionData) are located.}

\item{layout}{By default \code{'layout.fruchterman.reingold'}. It can be set 
to any other of the possible igraph layouts.}

\item{selectValue}{By default \code{"score"} variable will be selected. Change
it to any of the other possible variables (\code{'fdr'},\code{'odds ratio'}, 
\code{'phi'}, \code{'rr'}).}

\item{title}{Determines the title of the network figure. By default 
\code{'Comorbidity network'}.}

\item{cutOff}{By default \code{'0.05'}. The value of the argument can be changed 
to any other numeric variable, according to the range of the selected value.}

\item{npairs}{by default \code{'0'}.  The value of the argument can be changed
to any other numeric variable to show in the network only those comorbidities 
suffered by at least \code{npairs} of patients.}

\item{prop}{Determines the node size proportionality. By default it is set to 
\code{1}. The value of the argument can be changed to any other numeric variable.}

\item{interactive}{Determines if the output network is interactive or not. 
By default the \code{interactive} argument is set up as \code{FALSE}. The value 
of the argument can be changed to \code{TRUE}, as a result an interactive 
network will be obtained.}

\item{diseaseColor}{Determines the node color for the disorder of interest.
By default it is set to \code{"pink"}.}

\item{comorColor}{Determines the node color for the comorbid disorders. 
By default it is set to \code{"lightblue"}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A network
}
\description{
Given an object of class \code{cAnalysis} or \code{molecularcAnalysis} obtained from a comorbidity analysis, 
a network is obtained.
}
\examples{
load(system.file("extdata", "comorMale.RData", package="comoRbidity"))
comorbidityNetwork <- network( 
              input            = comorMale,
              databasePth      = system.file("extdata", package="comoRbidity"), 
              selectValue      = "score",
              cutOff           = 1.5,
              npairs           = 2,
              title            = "Example comorbidity network"
              )
}

