% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhds2010}
\alias{nhds2010}
\title{Adult same-day discharges, 2010}
\format{
A data frame with 2,210 rows and 15 variables:
\describe{
\item{ageu}{Units for age}
\item{age}{Age}
\item{sex}{Sex}
\item{race}{Race}
\item{month}{Discharge month}
\item{status}{Discharge status}
\item{region}{Region}
\item{atype}{Type of admission}
\item{dx1}{Diagnosis 1, ICD9-CM}
\item{dx2}{Diagnosis 2, ICD9-CM}
\item{dx3}{Diagnosis 3, ICD9-CM, imported incorrectly}
\item{dx3corr}{Diagnosis 3, ICD9-CM, corrected}
\item{pr1}{Procedure 1}
\item{wgt}{Frequency weight}
\item{recid}{Order of record (raw data)}
}
}
\usage{
nhds2010
}
\description{
A dataset containing adult same-day discharges from 2010, obtained from Stata 17.
}
\note{
The R code used to download and process the dataset from Stata is available \href{https://raw.githubusercontent.com/ellessenne/comorbidity/master/data-raw/make-data.R}{here}.
}
\keyword{datasets}
