% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationAge.R
\name{populationAge}
\alias{populationAge}
\title{Age distribution comparison: patients subset vs all population.}
\usage{
populationAge(input, codesPth, databasePth, type = "together",
  disorderColor = "gold", allColor = "darkblue", interactive = FALSE,
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{A comorbidity object, obtained with the query function.}

\item{codesPth}{Determines the path where the file with the index diseases is 
located (indexDiseaseCode)}

\item{databasePth}{Determines the path where the three required input files 
(patientData, diagnosisData, admissionData) are located.}

\item{type}{by Befault \code{'together'}. This argument allows the user to 
select the output barplot type. The other possible option is \code
{'separate'}}

\item{disorderColor}{Determines the bar color representing the disorder patients. 
By default \code{gold}.}

\item{allColor}{Determines the bar color representing the whole population. 
By default \code{darkblue}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A barplot comparing the ages distributions
}
\description{
Given an object of class \code{comoRbidity} the age distribution of your patients
of interest (those that have at least one code of the ones contained in indexDiseaseCodes.txt 
file) is represent comparing with the age distribution of the your whole patients 
database.
}
\examples{
load(system.file("extdata", "comorbidity.RData", package="comoRbidity"))
popAge <- populationAge( input       = comor_obj, 
                         codesPth    = system.file("extdata", package="comoRbidity"), 
                         databasePth = system.file("extdata", package="comoRbidity"),
                         type        = "separate"
              )
}
