% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query your data and generates a \code{comoRbidity}}
\usage{
query(databasePth, codesPth, birthDataSep, admissionDataSep,
  intraCodes = FALSE, aggregatedDis = FALSE, python = FALSE,
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{databasePth}{Determines the path where the three required input files 
(patientData, diagnosisData, admissionData) are located.}

\item{codesPth}{Determines the path where the file with the index diseases is 
located (indexDiseaseCode)}

\item{birthDataSep}{Determines the separator symbol used in the birth date.}

\item{admissionDataSep}{Determines the separator symbol used in the 
admission date.}

\item{intraCodes}{Comorbidities will be estimated only between the index 
diseases. By default \code{FALSE}}

\item{aggregatedDis}{Data extraction is done using the Agg column from the 
index disease file,  that collapse the diseases in a supperior class.
By default \code{FALSE}}

\item{python}{By default \code{FALSE}. Change it to \code{TRUE} to run the 
query quicklier using python. It is necessary to have python installed.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{comoRbidity}
}
\description{
Given 4 files (patientData.txt, diagnosisData.txt, admissionData.txt and 
indexDiseaseCodes.txt), generates some RData and create an object of type 
\code{comoRbidity}.
}
\examples{
ex1 <- query( databasePth      = system.file("extdata", package="comoRbidity"),
              codesPth         = system.file("extdata", package="comoRbidity"),
              birthDataSep     = "-",
              admissionDataSep = "-",
              python           = FALSE)
}
