% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_diag.R
\name{sample_diag}
\alias{sample_diag}
\title{Generate pseudo-ICD-10 diagnoses.}
\usage{
sample_diag(n = 1)
}
\arguments{
\item{n}{Number of pseudo-ICD-10 codes to generate.}
}
\value{
A vector of \code{n} pseudo-ICD-10 diagnostic codes.
}
\description{
A simple function to simulate pseudo-ICD-10 diagnostic codes at random.
}
\examples{
# Simulate 10 ICD-10 codes
sample_diag(10)

# Simulate a tidy dataset with 15 individuals and 200 rows
set.seed(1)
x <- data.frame(
  id = sample(1:15, size = 200, replace = TRUE),
  code = sample_diag(200),
  stringsAsFactors = FALSE)
head(x)

}
