% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_year_quarter.R
\name{sample_year_quarter}
\alias{sample_year_quarter}
\title{Simulate Year and Quarter}
\usage{
sample_year_quarter(n = 1)
}
\arguments{
\item{n}{Number of Year/Quarter pairs to generate.}
}
\value{
A data.frame of with \code{n} years and quarters.
}
\description{
A simple function to simulate Year and Quarter data at random.
}
\examples{
# Simulate 10 Year/Quarter pairs
sample_year_quarter(10)

# Simulate a tidy dataset with 15 individuals and 200 rows
set.seed(1)
x <- data.frame(
  id = sample(1:15, size = 200, replace = TRUE),
  stringsAsFactors = FALSE
)
x[c('Year', 'Quarter')] = sample_year_quarter(15)[x$id,]
head(x)
}
