% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexRatio.R
\name{sexRatio}
\alias{sexRatio}
\title{Sex Ratio Analysis.}
\usage{
sexRatio(female, male, fisherTest = 0, fisherCutOff = 0.05,
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{female}{A \code{comoRbidity} object, containing the comorbidity data for
female patients.}

\item{male}{A \code{comoRbidity} object, containing the comorbidity data for
male patients.}

\item{fisherTest}{By default the fisher test is not performed. The fisherTest 
default argument is set to \code{0}. It can be changed to 1 in order to perform 
the fisher test to each common comorbidity present in male and female.}

\item{fisherCutOff}{by default \code{0.05}. The value of the argument can be changed 
to any other numeric variable, according to the range of the fisher value.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A dataframe with the sex ratio estimation is obtained
}
\description{
Given two objects of class \code{comoRbidity}, one for each gender, the sex 
ratio estimation is performed.
}
\examples{
load(system.file("extdata", "comorMale.RData", package="comoRbidity"))
load(system.file("extdata", "comorFemale.RData", package="comoRbidity"))
srEstimation <- sexRatio( female       = comorFemale, 
                          male         = comorMale, 
                          fisherTest    = 0
              )
}

