% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryDiseases.R
\name{summaryDiseases}
\alias{summaryDiseases}
\title{A graphical summary of the diseases of interest.}
\usage{
summaryDiseases(input, database = "CURATED", type = "dis_barplot",
  assocGeneColor = "#E69F00", assocDiseaseColor = "#136593",
  dsiColor = "red", dpiColor = "black", interactive = FALSE,
  verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{input}{Object of \code{molecularComorbidity} class.}

\item{database}{The default value is set to \code{'CURATED'}. User can select 
any of the databases available in DisGeNET.}

\item{type}{Determines if the analysis is performed for genes or diseases. By 
default the argument value is set to \code{dis_barplot}, allowing to analyze 
the number of genes associated to each one of the index diseases as well as
the number of disorders that share some genes with them. It can be changed to 
\code{gene_barplot} to analyze and characterize the genes associated to the 
index diseases.}

\item{assocGeneColor}{Determines the bar color that represents the number of associated genes. 
By default it is set to "#E69F00".}

\item{assocDiseaseColor}{Determines the bar color that represents the number of associated diseases. 
By default it is set to "#136593".}

\item{dsiColor}{Determines the dot color that represents the Disease Specificity Index (DSI). 
By default it is set to "red".}

\item{dpiColor}{Determines the dot color that represents the Disease Pleiotropy Index (DPI). 
By default it is set to "black".}

\item{interactive}{Determines if the output barplot is interactive or not. 
By default the \code{interactive} argument is set up as \code{FALSE}. The value 
of the argument can be changed to \code{TRUE}, as a result an interactive 
barplot generated with Shiny will be obtained. This argument is only available 
when \code{gene_barplot} type is selected.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A barplot characterizing the genes or the disease in terms of gene-disease
association.
}
\description{
Given an object of class \code{molecularComorbidity}, a graphical summary of 
the main characteristic of the disorders is obtained.
}
\examples{
load(system.file("extdata", "mc.RData", package="comoRbidity"))
summaryDiseases( input    = mc, 
                 type     = "dis_barplot",
                 database = "CURATED" 
           )
}
