% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlson.R
\name{charlson}
\alias{charlson}
\title{Charlson Comorbidity Score.}
\usage{
charlson(x, id, code, assign0 = TRUE, factorise = FALSE, labelled = TRUE,
  tidy.codes = TRUE, parallel = FALSE, mc.cores = parallel::detectCores())
}
\arguments{
\item{x}{A tidy data frame with one column containing an individual ID and a column containing all ICD-10 codes.}

\item{id}{Column of \code{x} containing the individual ID.}

\item{code}{Column of \code{x} containing ICD-10 codes. Codes must be in upper case with no puctuation in order to be properly recognised.}

\item{assign0}{Apply a hierarchy of comorbities. If \code{TRUE}, should a comorbidity be present in a patient with different degrees of severity, then the milder form will be assigned to 0 and therefore not counted. By doing this, a type of comorbidity is not counted more than once in each patient.}

\item{factorise}{Return comorbidities as factors rather than numeric, where (1 = presence of comorbidity, 0 = otherwise). Defaults to \code{FALSE}.}

\item{labelled}{Attach labels to each comorbidity, compatible with the RStudio viewer via the \code{\link[utils:View]{utils::View()}} function. Defaults to \code{TRUE}.}

\item{tidy.codes}{Tidy ICD-10 codes? If \code{TRUE}, all codes are converted to upper case and all punctuation is stipped (REGEXPR: \code{[[:punct:]]}). Defaults to \code{TRUE}.

[[:punct:]: R:[:punct:}

\item{parallel}{Run the computation in parallel? See the \code{Notes} section for more information. Defaults to \code{FALSE}.}

\item{mc.cores}{The number of cores to use when running the computations in parallel. Defaults to all available cores.}
}
\value{
A data frame with \code{id}, columns relative to each Charlson domain, weighted Charlson Score, and grouped Charlson Index, with one row per individual. The following variables are included in the dataset:
\itemize{
\item The \code{id} variable as defined by the user;
\item \code{ami}, for acute myocardial infarction;
\item \code{chf}, for congestive heart failure;
\item \code{pvd}, for peripheral vascular disease;
\item \code{cevd}, for cerebrovascular disease;
\item \code{dementia}, for dementia;
\item \code{copd}, chronic obstructive pulmonary disease;
\item \code{rheumd}, for rheumatoid disease;
\item \code{pud}, for peptic ulcer disease;
\item \code{mld}, for mild liver disease;
\item \code{diab}, for diabetes without complications;
\item \code{diabwc}, for diabetes with complications;
\item \code{hp}, for hemiplegia or paraplegia;
\item \code{rend}, for renal disease;
\item \code{canc}, for cancer (any malignancy);
\item \code{msld}, for moderate or severe liver disease;
\item \code{metacanc}, for metastatic solid tumor;
\item \code{aids}, for AIDS/HIV;
\item \code{score}, for the weighted Charlson score;
\item \code{index}, for the grouped Charlson index.
}

Labels are presented to the user when using the RStudio viewer (e.g. via the \code{\link[utils:View]{utils::View()}} function) for convenience.
}
\description{
Computes the weighted Charlson Score, using ICD-10 codes.
}
\details{
This function is based on the ICD-10 Charlson Score definition proposed by Quan \emph{et al.} in 2005. ICD-10 codes must be in upper case and with no punctuation in order to be properly recognised; set \code{tidy.codes = TRUE} to properly tidy the codes automatically.
To run the calculations in parallel set \code{parallel = TRUE}. This is based on \code{\link[parallel:parLapply]{parallel::parLapply()}}, and it is possible to set the number of cores to use via the \code{mc.cores} argument, which defaults to using all the cores available.
}
\examples{
set.seed(1)
x <- data.frame(
  id = sample(1:15, size = 200, replace = TRUE),
  code = sample_diag(200),
  stringsAsFactors = FALSE)

charlson(x = x, id = "id", code = "code")

}
\references{
Quan H, Sundararajan V, Halfon P, Fong A, Burnand B, Luthi J-C, et al. Coding Algorithms for Defining Comorbidities in ICD-9-CM and ICD-10 Administrative Data. Medical Care 2005; 43:1130-1139.
}
