% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionality.R
\name{directionality}
\alias{directionality}
\title{Query your data and generates a \code{datafame}}
\usage{
directionality(input, databasePth, minPairs = 1, gender, ageRange = c(0,
  100), days, dataSep = "-", verbose)
}
\arguments{
\item{input}{A \code{cAnalysis} object, containing the comorbidity data for
a set of patients.}

\item{databasePth}{Determines the path where the intermediate RData objects 
have been created. It is the same path where the three required input files 
(patientData, diagnosisData, admissionData) are located.}

\item{minPairs}{Determines the minimum number of patients that must suffer the
comorbidity. By default the \code{minPairs} value is set to \code{1}. The value 
of the argument can be changed to any other numeric variable.}

\item{gender}{Determine what is the gender of interest for 
performing the comorbidity analysis.}

\item{ageRange}{Determines what is the age range of interest for
performing the comorbidity analysis. By default it is set from 0 to 100 
years old.}

\item{days}{Determines the number of days of difference needed for considering 
two diseases as comorbid.}

\item{dataSep}{Determines the separator symbol used in the admission date.
By default \code{"-"}.}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{dataframe}
}
\description{
Given an object of class \code{cAnalysis} obtained from a comorbidity analysis,
the comorbidity direction can be estimated. An object of class \code{dataframe} 
is obtained.
}
\examples{
load(system.file("extdata", "comorMale.RData", package="comoRbidity"))
directionEx <- directionality( input       = comorMale, 
                               databasePth    = system.file("extdata", package="comoRbidity"), 
                               minPairs         = 1, 
                               gender           = "Male",
                               ageRange         = c(0,80),
                               days             = "0",
                               dataSep          = "-"
              )
}

