% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_diag.R
\name{sample_diag}
\alias{sample_diag}
\title{Generate pseudo-ICD-10 diagnoses.}
\usage{
sample_diag(n = 1)
}
\arguments{
\item{n}{Number of pseudo-ICD-10 codes to generate.}
}
\value{
A vector of \code{n} pseudo-ICD-10 diagnoses.
}
\description{
A simple random generator for pseudo-ICD-10 diagnoses.
}
\examples{
sample_diag(10)

}
\author{
Alessandro Gasparini, \email{alessandro.gasparini@ki.se}
}

